### 2022-06-10 #542 @s3alfisc
fwildclusterboot presubmission
Submitting Author Name: Alexander Fischer
Submitting Author Github Handle: @s3alfisc
Other Package Authors Github handles: @droodman,
Repository:  https://github.com/s3alfisc/fwildclusterboot
Submission type: Pre-submission

---

```
Description: Implementation of the fast algorithm for wild cluster bootstrap
             inference developed in Roodman et al (2019, STATA Journal) for
             linear regression models <doi:10.1177/1536867X19830877>,
             which makes it feasible to quickly calculate bootstrap test
             statistics based on a large number of bootstrap draws even for
             large samples. Multiway clustering, regression weights,
             bootstrap weights, fixed effects and subcluster bootstrapping
             are supported. Further, both restricted (WCR) and unrestricted
             (WCU) bootstrap are supported. Methods are provided for a variety
             of fitted models, including 'lm()', 'feols()'
             (from package 'fixest') and 'felm()' (from package 'lfe').
             Additionally implements a heteroskedasticity-robust (HC1) wild
             bootstrap.
             Further, the package provides an R binding to 'WildBootTests.jl',
             which provides additional speed gains and functionality,
             including the 'WRE' bootstrap for instrumental variable models
             (based on models of type 'ivreg()' from package 'ivreg')
             and hypotheses with q > 1.
```

	- [ ] text data

`fwildclusterboot` conducts inference for (linear) regression models via a wild (cluster) bootstrap. It further serves as an R binding of the [WildBootTests.jl](https://github.com/droodman/WildBootTests.jl) library.

Yes, I have worked with the `srr` package and have a [draft ](https://github.com/s3alfisc/fwildclusterboot/blob/ropensci/R/srr-stats-standards.R)available (but it is currently not in the main branch).

The target audience is academic social scientists (economics, political science, sociology). `fwildclusterboot` should be used whenever regression errors are "clustered" into few groups, in which case inference based on asymptotic approximations might fail.

Other R packages that implement the wild cluster bootstrap are `sandwich` via its `vcovBS` function and the `clusterSEs` package. `fwildclusterboot` implements a [significantly faster algorithm](https://www.r-bloggers.com/2021/12/1000x-faster-wild-cluster-bootstrap-inference-in-r-with-fwildclusterboot-%f0%9f%9a%80/). Furter, `fwildclusterboot` offers additional functionality, e.g. the subcluster bootstrap. Through `WildBootTests.jl`, it also allows to run a highly optimized version of the WRE bootstrap for IV regressions (Davidson & MacKinnon, 2010) , which is not available in any other R package.

Yes.

`fwildclusterboot` implements the "fast" wild cluster bootstrap in R, but also allows to call `WildBootTests.jl` via the `JuliaConnectoR` package. It's therefore (also) a wrapper package, and you might consider it to be out of scope?
