+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   WildBootTests: WildBootTests.html
   fwildclusterboot: fwildclusterboot.html
-last_built: 2022-06-10T21:06Z
+last_built: 2022-06-11T09:06Z
 urls:
   reference: https://s3alfisc.github.io/fwildclusterboot/reference
   article: https://s3alfisc.github.io/fwildclusterboot/articles
