### 2022-06-11 78d48e9
add autotest checks
