+++ DESCRIPTION
@@ -79,7 +79,8 @@ Suggests:
     modelsummary,
     bench,
     testthat (>= 3.0.0), 
-    tibble
+    tibble, 
+    autotest
 Encoding: UTF-8
 LazyData: true
 Roxygen: list(markdown = TRUE)
