### 2022-06-11 6b3cb15
first draft of using srr package
