### 2022-06-11 a41573c
add nobs, confint, pval, teststat methods
