### 2022-06-11 dcbb845
document non-exported functions
