+++ R/preprocess2.R
@@ -1,6 +1,4 @@
-#' Fast wild cluster bootstrap inference
-#'
-#'
+
 #' `preprocess2` is a S3 method that fetches data from several model
 #' objectects for use with `boottest()`.
 #'
@@ -16,7 +14,19 @@ preprocess2 <- function(object, ...) {
 
 
 preprocess2.fixest <- function(object, clustid, R, param, fe, boot_algo, bootcluster){
-
+  
+  #' preprocess data for objects of type fixest
+  #' 
+  #' @param object an object of type fixest
+  #' @param clustid a character string containing the name(s) of the cluster variables
+  #' @param R Hypothesis Vector giving linear combinations of coefficients. 
+  #' @param fe character vector. name of the fixed effect to be projected out in the bootstrap
+  #' @param param character vector. names of the parameter(s) to test
+  #' @param boot_algo The bootstrap algorithm to run. Either "R" or "WildBootTests.jl"
+  #' @param bootcluster a character string containing the name(s) of the bootcluster variables. Alternatively, "min" or "max"
+  #'
+  #' @noRd
+  #'  
   #' @method preprocess2 fixest
 
   call <- object$call
@@ -179,6 +189,18 @@ preprocess2.fixest <- function(object, clustid, R, param, fe, boot_algo, bootclu
 
 preprocess2.felm <- function(object, clustid, R, param, fe, boot_algo, bootcluster){
 
+  #' preprocess data for objects of type felm
+  #' 
+  #' @param object an object of type felm
+  #' @param clustid a character string containing the name(s) of the cluster variables
+  #' @param R Hypothesis Vector giving linear combinations of coefficients.  #' @param fe 
+  #' @param fe character vector. name of the fixed effect to be projected out in the bootstrap
+  #' @param param character vector. names of the parameter(s) to test
+  #' @param boot_algo The bootstrap algorithm to run. Either "R" or "WildBootTests.jl"
+  #' @param bootcluster a character string containing the name(s) of the bootcluster variables. Alternatively, "min" or "max"
+  #' 
+  #' @noRd
+  #' 
   #' @method preprocess2 felm
 
   call <- object$call
@@ -323,8 +345,19 @@ preprocess2.felm <- function(object, clustid, R, param, fe, boot_algo, bootclust
 
 preprocess2.lm <- function(object, clustid, R, param, boot_algo, bootcluster){
 
+  #' preprocess data for objects of type lm
+  #' 
+  #' @param object an object of type lm
+  #' @param clustid a character string containing the name(s) of the cluster variables
+  #' @param R Hypothesis Vector giving linear combinations of coefficients.  #' @param fe 
+  #' @param param character vector. names of the parameter(s) to test
+  #' @param boot_algo The bootstrap algorithm to run. Either "R" or "WildBootTests.jl"
+  #' @param bootcluster a character string containing the name(s) of the bootcluster variables. Alternatively, "min" or "max"
+  #' 
+  #' @noRd
+  #' 
   #' @method preprocess2 lm
-
+  
   call <- object$call
   call_env <- environment(formula(object))
   fml <- formula(object)
@@ -428,8 +461,19 @@ preprocess2.lm <- function(object, clustid, R, param, boot_algo, bootcluster){
 
 preprocess2.ivreg <- function(object, clustid, R, param, boot_algo, bootcluster){
 
+  #' preprocess data for objects of type ivreg
+  #' 
+  #' @param object an object of type ivreg
+  #' @param clustid a character string containing the name(s) of the cluster variables
+  #' @param R Hypothesis Vector giving linear combinations of coefficients.  #' @param fe 
+  #' @param param character vector. names of the parameter(s) to test
+  #' @param boot_algo The bootstrap algorithm to run. Either "R" or "WildBootTests.jl"
[…]
