### 2022-06-11 1eeb783
add barebone support for plm() with pooling
