### 2022-06-12 e839855
add definition of statistical terms and definitions in extra vignette
