### 2022-06-12 #542 @emilyriederer
Hi @s3alfisc - thanks so much for submitting your package. I especially appreciate all of the details (and impressive benchmarking results!) in the best-in-class answer.

As a general matter, this seems in-scope for a regression package. Based on your work with the statistical standards, could you please comment on whether you believe that the package is on track to meet at least half of the general + category-specific standards?

Thanks also for the call-out on the optional functionality to call the Julia implementation. We are also planning, but do not yet have standards, for a statistical wrapper package. A member of the statistics peer review team may comment further on whether or not this package could fit that category also.
