### 2022-06-13 #542 @s3alfisc
Hi @emilyriederer , thanks for your feedback! I have uploaded my comments based on the statistical software roclets in a separate branch [here ](https://github.com/s3alfisc/fwildclusterboot/blob/ropensci/R/srr-stats-standards.R).
