### 2022-06-17 #542 @mpadge
@s3alfisc @emilyriederer I've had a look through the code, and do not think this package should really be considered a statistical "wrapper" package, as it only constucts a single external call to one Julia package. The Julia connection is entirely optional for package functionality, and in terms of code and algorithms represents only a very small portion of the code. I suggest the review process can proceed under the single category nominated above.

@s3alfisc I note that your current version documents compliance with 59 / 115 standards, which is > 50%, so okay to proceed. The `srr_report()` nevertheless notes standard G2.15 appears to be missing. Could you please check and rectify if possible? Note also that our automated check system currently works on GitHub default branch only, but we're happy to use your submission to develop an appropriate workflow for non-default branches. Until then, the checks might have to be generated for default branch, after which I'll manually remove that comment and re-generate them for your "ropensci" branch. Thanks!
