### 2022-06-17 #542 @s3alfisc
Hi @mpadge , thanks for your feedback! I will spend some time cleaning up the package over the next days (documenting all srr roclets, add G.2.15, and merge everything into the main branch) and then I will submit `fwildclusterboot`! =)
