### 2022-06-17 #542 @mpadge
@s3alfisc No need to merge if you'd rather not. We do want our system to one day work on non-default branches, so as said are happy to use your submission to test that, if that's easier for you. That said, we do generally advise against this, because then you'll be stuck implementing changes to reviews in your non-default branch, which may make your own workflow less robust. Up to you.
